+++ src/vanilla/utils.ts
@@ -3,9 +3,9 @@ export { watch } from './utils/watch.ts'
 export { devtools } from './utils/devtools.ts'
 export { deepClone } from './utils/deepClone.ts'
 export { proxySet } from './utils/proxySet.ts'
-export { proxyMap } from './utils/proxyMap-indexMap-keyvals.ts'
+export { proxyMap } from './utils/proxyMap-old.ts'
 
 export { proxyMap as newProxyMap } from './utils/proxyMap-indexMap-filled.ts'
 export { proxyMap as newProxyMapKeyVals } from './utils/proxyMap-indexMap-keyvals.ts'
-export { proxyMap as newProxyMapRawMap } from './utils/proxyMap-rawMap.ts'
+export { proxyMap as newProxyMapRawMap1 } from './utils/proxyMap-rawMap1.ts'
 export { proxyMap as newProxyMapTree1 } from './utils/proxyMap-tree1.ts'
